+++ RELEASE_NOTES.md
@@ -44,6 +44,7 @@ By deprecating these functions, we can simplify the library in 1.0, and focus mo
 * `RichTextAttributeReader` has deprecated a lot of functions.
 * `RichTextAttributeWriter` has deprecated a lot of functions.
 * `RichTextCoordinator` functions that simply triggered `handle(_:)` have been deprecated.
+* `RTKL10n.bundle` has been deprecated, since we can just use the `.module` bundle from now.